backend/requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
Django==5.1.6
djangorestframework==3.15.2
drf-yasg==1.21.8
geographiclib==2.0
geopy==2.4.1
gunicorn==23.0.0
h11==0.14.0
h3==4.2.1
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
numpy==2.2.3
packaging==24.2
pluggy==1.5.0
pycparser==2.22
pytest==8.3.4
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
timezonefinder==6.5.8
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
whitenoise==6.9.0
